I created DBEngine.cpp & DBEngine.h which is one of the examples for Creating Character instances and Item instances(WeaponItem and ArmorItem classes) from XML file.
Eventually, the functionality of this class could be added to GameEngine class. Now I just didn't want to mess up with GameEngine class, so, I just crated a seperate class.

You guys can comment to this class freely.

I tested this class several time with various XMLs and it worked perfectly fine for me.

Which means :

* Create Character instance with its name, level, weapon and armor.
* Create Item instance with its name and attribute.
* Character can equip items now.