Updated so different levels actually work. (play through the first one and see the cool second one)
Also added event key recording and playback support, which seems to be working but I want to investigate a bit more before declaring it 100% working. To use, launch the game with the parameter -record filename (where filename is the name of the file you want to record to, a txt file basically) or use the parameter -playback filename (where filename is a file that exists and has previously-recorded key presses)

The keys are written to the file for every press, so if the game crashes while recording you'll have all the keypresses up until then. Haven't figured out a good way of integrating the different command line options in Code::Blocks yet...

Also the AI doesn't pause in playback mode.