Modified wayyy too much for one go, sorry guys. Anyways went on a const and referenced craze, so things should be safer and more efficient now, well the stuff I changed that is.
Also I took out some usage of using namespace std;, since this is bad in practice and we should be doing std::string for strings, etc. We can cover up our own usage of vectors and stuff by making typedefs and then the std::vector<blah> is only in the typedef definition, and we can only type BlahVec where-ever we want (not std::BlahVec or std::vector<Blah>)

Also added the actual enum names in the big switch case in UIGrid. 

Please let me know if there are any issues after this commit, I found a few bugs in the process of doing it and fixed them.