Fixed a incorrectly-ended comment that screwed compilation, sorry about that. Also updated the project file since many headers we are using weren't included in the project.
Please, we are using the project to view files that are in the game, so add headers to the project if you are using them! They aren't compiled, but at least we know which ones are used, and we are supposed to open files in the project's file explorer not manually going open file-> and then locating it. Also C::B automatically checks if any of the files in the project has changed and will recompile what is necessary, which prevents you from having to re-save or force re-compilation of files not included in the project file, but are used in the project.