Big commit.
Most significant change is that Party Status Layout now shows the status of the 4 members of the party.

Also the Weapon and Armor lists created in the layout work, but I have them commented out because all of the functionality is not there yet.

Market is complete (Armory will soon follow)

I made changed to Item, ArmorItem, and WeaponItem to now include their 'type' amongst the attributes (I had difficulty differentiating between Armor and Weapons in a list of Item pointers)

New class: UIItemWindow, though it may seem extraneous, will make "equipping" functionality much easier to implement (and other functionality as well)

Adjustments to the OverMap elements was to update the GameEngine as to what city is being selected dynamically and adding a GetCity method to the 'tile' class

Added GetCursorPos method to UIMenu, and also ensure the mCursorFunc is deleted in the UIMenu destructor.

Added a couple get/set methods to GameEngine (ignore the comments change... I explain in the inline comments if you want to know)

..and some new pics:
www.sfu.ca/~aco1/images/menu_back_small.png

and I don't know if I mentioned this before, but I updated the credits pic:
www.sfu.ca/~aco1/images/credits.png