Updated UILoseLayout to better backgound, adjusted position of text and changed some information (relevant to the current version)
Made 2 changes with PartyStatusLayout.  When in battle, the Equip-operations appear ghosted (not selectable).  Also made it so that when your 'inventory' is empty, no menu appears and there's text telling you that you have no weapon/armor to equip.  This text does not appear when you re-enter the screen.

With UIMenu just added a method "Empty", which returns whether the menu is empty or not.